backend/requirements.txt
fastapi
uvicorn
requests
pandas
databricks-sql-connector
prometheus-fastapi-instrumentator
prometheus-client
python-multipart
mlflow

# Add dbt support for Databricks
dbt-core
dbt-databricks
